% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_boosting_fit.R
\name{group_boosting_fit}
\alias{group_boosting_fit}
\title{Group boosting fitting}
\usage{
group_boosting_fit(
  X,
  Y,
  group,
  total_steps = 50000,
  step_size = 1e-06,
  adj_var = 999,
  stop_tol = -1e-07,
  gamma = 0,
  lasso_lambda = 0.0314,
  weighted = FALSE
)
}
\arguments{
\item{X}{matrix of predictors including any adjustment variables, which may not be grouped, of dimension n $/times$ p.}

\item{Y}{vector of the response variable, should be length n.}

\item{group}{vector containing an indicator of group membership of the p predictors in \code{X}}

\item{total_steps}{integer value of the maximum number of iterations to run the boosting algorithm}

\item{step_size}{a double value used to control how much to increase the estimate by at each step}

\item{adj_var}{integer number for the number of adjustment variables, 999 used to denote no adjustment variables}

\item{stop_tol}{double value used to determine what change in AIC/BIC/EBIC is small enough to stop iterating}

\item{gamma}{double value used to determine the penalty when using EBIC as stopping criteria}

\item{lasso_lambda}{double value of lambda used for lasso if regressing out the effect of adjustment variables}

\item{weighted}{default value is FALSE, set to TRUE if you want to use a weighted version to summarize within groups for the first stage}
}
\value{
final_beta contains a vector with the final coefficient estimates, update1 is a vector that contains the coefficients from the first stage estimation, AIC1 contains the AIC path of the iterations from the first stage update, and BIC2 contains the BIC path from the second stage iterations
}
\description{
Group boosting fitting
}
