% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simul_dat_group}
\alias{simul_dat_group}
\title{Simulate grouped data}
\usage{
simul_dat_group(
  nodes = 25,
  n = 100,
  num.groups = 5,
  q.groups = 3,
  sparse_g = 0.1,
  dense_g = 0.9,
  effect_size = 1,
  SNR = 0.9,
  structure = "block",
  adj_var = 999,
  rho_adj_var,
  seed = 2020
)
}
\arguments{
\item{nodes}{integer value of the number of nodes in the network}

\item{n}{integer, number of subjects}

\item{num.groups}{integer, number of groups that the nodes in the network belong to}

\item{q.groups}{integer, number of group with non-zero signal}

\item{sparse_g}{ineger, number of groups with sparse signal}

\item{dense_g}{integer, number of groups with dense signal among the edges}

\item{effect_size}{double value indicating the effect size for the non-zero signals}

\item{SNR}{double value for the desired signal to noise ratio}

\item{structure}{default is "block" which defines the grouping among edges in the network, "nodes" is also available if the grouping should be by nodes instead}

\item{adj_var}{integer for the number of adjustment variables}

\item{rho_adj_var}{double value of the average correlation between connectivity of the edges and the adjustment variables}

\item{seed}{integer value to set.seed}
}
\value{
Y contains the simulated response vector, X is a matrix of predictors including any adjustment variables, and group is a vector of the group membership of each edg
}
\description{
Simulate grouped data
}
